requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
Flask-Login==0.6.3
Werkzeug==2.2.3
email-validator==2.3.0